Improve type safety and error handling across modules
Updated type annotations in imu/utils.py and models/gating.py for better clarity and type safety. Enhanced error handling in io/dlc.py for DLC file loading, including checks for MultiIndex columns and DataFrame types. Fixed a potential bug in io/spikegadgets.py by ensuring sampling_rate is always a float. Added assertion in runtime/online.py to guarantee EKF filter availability after initialization. In sim/synthetic.py, replaced mutable default arguments with field(default_factory=list) for dataclass lists and clarified heading type.

+++ src/trodestrack/io/spikegadgets.py
@@ -504,7 +504,7 @@ def load_spikegadgets_csv(file_path: Path) -> SpikeGadgetsIMUData:
     # Estimate sampling rate
     if len(timestamps) > 1:
         dt = np.median(np.diff(timestamps))
-        sampling_rate = 1.0 / dt
+        sampling_rate = float(1.0 / dt)
     else:
         sampling_rate = 30000.0  # Default assumption
